Added PSR-1 and PSR-2 compatibility

+++ phpmyfaq/admin/trans.edit.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Handle ajax requests for the interface translation tool
+ * Handle ajax requests for the interface translation tool.
  *
  * PHP Version 5.5
  *
@@ -9,25 +9,26 @@
  * obtain one at http://mozilla.org/MPL/2.0/.
  *
  * @category  phpMyFAQ
- * @package   Administration
+ *
  * @author    Anatoliy Belsky <ab@php.net>
  * @copyright 2009-2015 phpMyFAQ Team
  * @license   http://www.mozilla.org/MPL/2.0/ Mozilla Public License Version 2.0
+ *
  * @link      http://www.phpmyfaq.de
  * @since     2009-05-11
  */
-
 if (!defined('IS_VALID_PHPMYFAQ')) {
     $protocol = 'http';
-    if (isset($_SERVER['HTTPS']) && strtoupper($_SERVER['HTTPS']) === 'ON'){
+    if (isset($_SERVER['HTTPS']) && strtoupper($_SERVER['HTTPS']) === 'ON') {
         $protocol = 'https';
     }
-    header('Location: ' . $protocol . '://' . $_SERVER['HTTP_HOST'] . dirname($_SERVER['SCRIPT_NAME']));
+    header('Location: '.$protocol.'://'.$_SERVER['HTTP_HOST'].dirname($_SERVER['SCRIPT_NAME']));
     exit();
 }
 
-if (!$user->perm->checkRight($user->getUserId(), "edittranslation")) {
+if (!$user->perm->checkRight($user->getUserId(), 'edittranslation')) {
     echo $PMF_LANG['err_NotAuth'];
+
     return;
 }
 
@@ -36,49 +37,47 @@ $translateLang = PMF_Filter::filterInput(INPUT_GET, 'translang', FILTER_SANITIZE
 $page = PMF_Filter::filterInput(INPUT_GET, 'page', FILTER_VALIDATE_INT);
 $page = 1 > $page ? 1 : $page;
 
-if (empty($translateLang) || !file_exists(PMF_ROOT_DIR . "/lang/language_$translateLang.php")) {
-    header("Location: ?action=translist");
+if (empty($translateLang) || !file_exists(PMF_ROOT_DIR."/lang/language_$translateLang.php")) {
+    header('Location: ?action=translist');
 }
 
-$tt = new PMF_TransTool;
+$tt = new PMF_TransTool();
 
-/**
+/*
  * There are meanwhile over 600 language
  * vars and we won't to show them all
  * at once, so let's paginate.
  */
 $itemsPerPage = 32;
-if(!isset($_SESSION['trans'])) {
-    /**
+if (!isset($_SESSION['trans'])) {
+    /*
      * English is our exemplary language
      */
-    $_SESSION['trans']['leftVarsOnly']  = $tt->getVars(PMF_ROOT_DIR . "/lang/language_en.php");
-    $_SESSION['trans']['rightVarsOnly'] = $tt->getVars(PMF_ROOT_DIR . "/lang/language_$translateLang.php");
-    
+    $_SESSION['trans']['leftVarsOnly'] = $tt->getVars(PMF_ROOT_DIR.'/lang/language_en.php');
+    $_SESSION['trans']['rightVarsOnly'] = $tt->getVars(PMF_ROOT_DIR."/lang/language_$translateLang.php");
 }
 
-$leftVarsOnly   = array_slice($_SESSION['trans']['leftVarsOnly'], 
-                              ($page-1)*$itemsPerPage,
+$leftVarsOnly = array_slice($_SESSION['trans']['leftVarsOnly'],
+                              ($page - 1) * $itemsPerPage,
                               $itemsPerPage);
-$rightVarsOnly  = &$_SESSION['trans']['rightVarsOnly'];
-
+$rightVarsOnly = &$_SESSION['trans']['rightVarsOnly'];
 
 $options = array(
-    'baseUrl' => PMF_Link::getSystemRelativeUri('index.php') . '?' . str_replace('&', '&amp;', $_SERVER['QUERY_STRING']),
-    'total'   => count($_SESSION['trans']['leftVarsOnly']),
-    'perPage' => $itemsPerPage
+    'baseUrl' => PMF_Link::getSystemRelativeUri('index.php').'?'.str_replace('&', '&amp;', $_SERVER['QUERY_STRING']),
+    'total' => count($_SESSION['trans']['leftVarsOnly']),
+    'perPage' => $itemsPerPage,
 );
 
 $pagination = new PMF_Pagination($faqConfig, $options);
-$pageBar    = $pagination->render();
+$pageBar = $pagination->render();
 
-/**
+/*
  * These keys always exist as they are defined when creating translation.
  * We use these values to add the correct number of input boxes.
  * Left column will always have 2 boxes, right - 1 to 6+ boxes.
  */
-$leftNPlurals  = (int)$_SESSION['trans']['leftVarsOnly']['PMF_LANG[nplurals]'];
-$rightNPlurals = (int)$rightVarsOnly['PMF_LANG[nplurals]'];
+$leftNPlurals = (int) $_SESSION['trans']['leftVarsOnly']['PMF_LANG[nplurals]'];
+$rightNPlurals = (int) $rightVarsOnly['PMF_LANG[nplurals]'];
 
 printf(
     '<header class="row"><div class="col-lg-12"><h2 class="page-header"><i class="fa fa-wrench"></i> %s</h2></div></header>',
@@ -99,12 +98,12 @@ $NPluralsErrorReported = false;
                     <th>en</th>
                     <th><?php echo $translateLang ?></th>
                 </tr>
-        <?php while(list($key, $line) = each($leftVarsOnly)): ?>
+        <?php while (list($key, $line) = each($leftVarsOnly)): ?>
         <?php
     // These parameters are not real translations, so don't offer to translate them
     if ($tt->isKeyIgnorable($key)) {
         echo "<tr>\n";
-        echo "<td>" . $key . "</td>\n";
+        echo '<td>'.$key."</td>\n";
         echo '<td><input class="form-control" type="text" value="'.PMF_String::htmlspecialchars($line).'" disabled="disabled" /></td>'."\n";
         echo '<td><input class="form-control" type="text" name="'.$key.'" value="'.PMF_String::htmlspecialchars($rightVarsOnly[$key]).'" disabled="disabled" />';
         echo '<input type="hidden" name="'.$key.'" value="'.PMF_String::htmlspecialchars($rightVarsOnly[$key]).'" /></td>'."\n";
@@ -112,13 +111,14 @@ $NPluralsErrorReported = false;
         continue;
     }
 
-    /**
+    /*
      *  Plural form support in translation interface
      */
 
     // We deal with the second plural form when dealing with the first, so skip it here
-    if ($tt->isKeyASecondPluralForm($key))
+    if ($tt->isKeyASecondPluralForm($key)) {
         continue;
+    }
 
     if ($tt->isKeyAFirstPluralForm($key)) {
         if ($rightNPlurals == -1) {
@@ -131,46 +131,49 @@ $NPluralsErrorReported = false;
             }
             continue;
         }
-        /**
+        /*
          * We echo one box for English and one for other language
          * because other language will always have at least 1 form
          */
         echo "<tr>\n";
-        echo "<td>".$key."</td>\n";
+        echo '<td>'.$key."</td>\n";
         echo '<td><input class="form-control" type="text" value="'.PMF_String::htmlspecialchars($line).'" disabled="disabled" /></td>'."\n";
         if (array_key_exists($key, $rightVarsOnly) && ($line != $rightVarsOnly[$key] ||
-           $tt->isKeyIgnorable($key) || $tt->isValIgnorable($line)))
+           $tt->isKeyIgnorable($key) || $tt->isValIgnorable($line))) {
             echo '<td><input class="form-control" type="text" name="'.$key.'" value="'.PMF_String::htmlspecialchars($rightVarsOnly[$key]).'" /></td>'."\n";
-        else
+        } else {
             echo '<td><input style="width: 300px;border-color: red;" type="text" name="'.$key.'" value="'.PMF_String::htmlspecialchars($line).'" /></td>'."\n";
+        }
         echo "</tr>\n";
 
         // Add second English form and translation
         $key2 = str_replace('[0]', '[1]', $key);
         echo "<tr>\n";
-        echo "<td>".$key2."</td>\n";
+        echo '<td>'.$key2."</td>\n";
         echo '<td><input class="form-control" type="text" value="'.PMF_String::htmlspecialchars($leftVarsOnly[$key2]).'" disabled="disabled" /></td>'."\n";
         if ($rightNPlurals == 1) {
             // Other language has only one form
             echo '<td><input class="form-control" type="text" value="'.$PMF_LANG['msgTransToolLanguageOnePlural'].'" disabled="disabled" /></td>'."\n";
         } else {
-            if (array_key_exists($key2, $rightVarsOnly))
+            if (array_key_exists($key2, $rightVarsOnly)) {
                 echo '<td><input class="form-control" type="text" name="'.$key2.'" value="'.PMF_String::htmlspecialchars($rightVarsOnly[$key2]).'" /></td>'."\n";
-            else
+            } else {
                 echo '<td><input class="form-control alert-danger" type="text" name="'.$key2.'" value="'.PMF_String::htmlspecialchars($leftVarsOnly[$key2]).'" /></td>'."\n";
+            }
         }
         echo "</tr>\n";
 
         // Other language has more than 2 forms
-        for ($i = 2; $i < $rightNPlurals; $i++) {
+        for ($i = 2; $i < $rightNPlurals; ++$i) {
             $keyI = str_replace('[0]', "[$i]", $key);
             echo "<tr>\n";
-            echo "<td>".$keyI."</td>\n";
+            echo '<td>'.$keyI."</td>\n";
             echo '<td><input class="form-control" type="text" value="" disabled="disabled" /></td>'."\n";
-            if (array_key_exists($keyI, $rightVarsOnly) && $leftVarsOnly[$key2] != $rightVarsOnly[$key])
+            if (array_key_exists($keyI, $rightVarsOnly) && $leftVarsOnly[$key2] != $rightVarsOnly[$key]) {
                 echo '<td><input class="form-control" type="text" name="'.$keyI.'" value="'.PMF_String::htmlspecialchars($rightVarsOnly[$keyI]).'" /></td>'."\n";
-            else
+            } else {
                 echo '<td><input class="form-control alert-danger" type="text" name="'.$keyI.'" value="'.PMF_String::htmlspecialchars($leftVarsOnly[$key2]).'" /></td>'."\n";
+            }
             echo "</tr>\n";
         }
         // We do not need to process this $key any further
@@ -200,7 +203,9 @@ $NPluralsErrorReported = false;
                             <?php echo $PMF_LANG['msgCancel'] ?>
                         </button>
                         <button class="btn btn-success" type="button"
-                                onclick="save()"<?php if (!is_writable(PMF_ROOT_DIR . "/lang/language_$translateLang.php")) { echo ' disabled="disabled"'; } ?>>
+                                onclick="save()"<?php if (!is_writable(PMF_ROOT_DIR."/lang/language_$translateLang.php")) {
+    echo ' disabled="disabled"';
+} ?>>
                             <?php echo $PMF_LANG['msgSave'] ?>
                         </button>
                     </td>
